requirement.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
PyJWT==1.7.1
django-cors-headers==3.5.0
django-dotenv==1.4.2
django-rest-swagger